smart-budget-agent/pyproject.toml
[project]
name = "smart-budget-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "firebase-admin (>=7.0.0,<8.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "google-api-python-client (>=2.176.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
